+++ src/main/java/com/mindata/ecserver/main/model/primary/EcContactEntity.java
@@ -10,7 +10,7 @@ import java.util.Date;
 @Entity
 @Table(name = "ec_contact_no_push")
 public class EcContactEntity {
-    private int id;
+    private Long id;
     /**
      * 人名
      */
@@ -72,7 +72,7 @@ public class EcContactEntity {
     /**
      * 对应爬虫的公司id
      */
-    private Integer compId;
+    private Long compId;
 
     private Date createTime;
 
@@ -85,21 +85,21 @@ public class EcContactEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id")
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
     @Basic
     @Column(name = "comp_id")
-    public Integer getCompId() {
+    public Long getCompId() {
         return compId;
     }
 
-    public void setCompId(Integer compId) {
+    public void setCompId(Long compId) {
         this.compId = compId;
     }
