+++ src/main/java/com/mindata/ecserver/main/model/secondary/PtUser.java
@@ -36,11 +36,11 @@ public class PtUser extends BaseEntity {
     /**
      * 公司ID
      */
-    private Integer companyId;
+    private Long companyId;
     /**
      * 部门ID
      */
-    private Integer departmentId;
+    private Long departmentId;
     /**
      * 称号
      */
@@ -113,11 +113,11 @@ public class PtUser extends BaseEntity {
         this.name = name;
     }
 
-    public Integer getCompanyId() {
+    public Long getCompanyId() {
         return companyId;
     }
 
-    public void setCompanyId(Integer companyId) {
+    public void setCompanyId(Long companyId) {
         this.companyId = companyId;
     }
 
@@ -137,11 +137,11 @@ public class PtUser extends BaseEntity {
         this.password = password;
     }
 
-    public Integer getDepartmentId() {
+    public Long getDepartmentId() {
         return departmentId;
     }
 
-    public void setDepartmentId(Integer departmentId) {
+    public void setDepartmentId(Long departmentId) {
         this.departmentId = departmentId;
     }
